Moved stategy name constants to the tests detector classes
