+++ strategies/changed/src/main/java/org/arquillian/smart/testing/vcs/git/ChangedFilesDetectorFactory.java
@@ -7,7 +7,7 @@ import org.arquillian.smart.testing.spi.StrategyConfiguration;
 import org.arquillian.smart.testing.spi.TestExecutionPlanner;
 import org.arquillian.smart.testing.spi.TestExecutionPlannerFactory;
 
-import static org.arquillian.smart.testing.vcs.git.ChangedConfiguration.CHANGED;
+import static org.arquillian.smart.testing.vcs.git.ChangedTestsDetector.CHANGED;
 
 public class ChangedFilesDetectorFactory implements TestExecutionPlannerFactory {
